refactor: 알람 생성 로직 위치 이동
