+++ app/src/main/java/woowacourse/movie/booking/complete/BookingCompleteContract.kt
@@ -5,11 +5,6 @@ import woowacourse.movie.ui.model.TicketUiModel
 interface BookingCompleteContract {
     interface View {
         fun showBookingCompleteResult(ticket: TicketUiModel)
-
-        fun makeAlarm(
-            ticket: TicketUiModel,
-            time: Long,
-        )
     }
 
     interface Presenter {
