+++ app/src/main/java/woowacourse/movie/booking/complete/BookingCompletePresenter.kt
@@ -1,15 +1,11 @@
 package woowacourse.movie.booking.complete
 
-import woowacourse.movie.data.SettingRepository
 import woowacourse.movie.mapper.toDomain
 import woowacourse.movie.model.Ticket
 import woowacourse.movie.ui.model.TicketUiModel
-import java.time.LocalDateTime
-import java.time.ZoneId
 
 class BookingCompletePresenter(
     private val view: BookingCompleteContract.View,
-    private val settingPreference: SettingRepository,
 ) : BookingCompleteContract.Presenter {
     private lateinit var ticket: Ticket
 
@@ -17,18 +13,5 @@ class BookingCompletePresenter(
         this.ticket = ticket.toDomain()
 
         view.showBookingCompleteResult(ticket)
-
-        if (settingPreference.isAlarmPermitted()) {
-            view.makeAlarm(ticket, calculateMovieAlarmTime())
-        }
-    }
-
-    private fun calculateMovieAlarmTime(): Long {
-        val dateTime = LocalDateTime.of(ticket.selectedDate, ticket.selectedTime)
-        return dateTime
-            .minusMinutes(30)
-            .atZone(ZoneId.systemDefault())
-            .toInstant()
-            .toEpochMilli()
     }
 }
